+++ tests/testthat/test-01-get_repos-GitHub.R
@@ -275,8 +275,8 @@ test_that("parse_search_response parses search response into repositories output
 test_that("parse_search_response parses search response into repositories output", {
   mockery::stub(
     github_testhost_priv$parse_search_response,
-    "api_engine$get_repos_from_org",
-    test_mocker$use("gh_repos_from_org")
+    "api_engine$get_repos",
+    test_mocker$use("gh_repos_by_ids")
   )
   gh_repos_by_code_table <- github_testhost_priv$parse_search_response(
     search_response = test_mocker$use("gh_search_repos_for_code"),
@@ -296,24 +296,6 @@ test_that("parse_search_response prints message", {
   )
   expect_snapshot(
     gh_repos_raw_output <- github_testhost_priv$parse_search_response(
-      search_response = test_mocker$use("gh_search_repos_for_code"),
-      org = gh_org,
-      output = "raw",
-      verbose = TRUE
-    )
-  )
-})
-
-test_that("parse_search_response turns to REST when first attempt returns error", {
-  internal_server_error <- list("repo" = list("id" = "1234"))
-  class(internal_server_error) <- c("graphql_error", "graphql_server_error")
-  mockery::stub(
-    github_testhost_priv$parse_search_response,
-    "api_engine$get_repos_from_org",
-    internal_server_error
-  )
-  expect_snapshot(
-    gh_repos_by_code_table <- github_testhost_priv$parse_search_response(
       search_response = test_mocker$use("gh_search_repos_for_code"),
       org = gh_org,
       output = "raw",
@@ -414,22 +396,6 @@ test_that("`get_repos_with_code_from_repos()` works", {
     )
   )
   expect_repos_table(repos_with_code_from_repos_full)
-})
-
-test_that("`get_repos_with_code_from_host()` pulls raw response", {
-  mockery::stub(
-    github_testhost_priv$get_repos_with_code_from_host,
-    "rest_engine$search_repos_for_code",
-    test_mocker$use("gh_search_repos_for_code")
-  )
-  expect_snapshot(
-    repos_with_code_from_host_raw <- github_testhost_priv$get_repos_with_code_from_host(
-      code = "shiny",
-      in_files = c("DESCRIPTION", "NAMESPACE"),
-      output = "raw",
-      verbose = TRUE
-    )
-  )
 })
 
 test_that("`get_repos_with_code_from_host()` pulls raw response", {
@@ -731,4 +697,13 @@ test_that("filtering by language works", {
     nrow(repos_table),
     nrow(filtered_repos)
   )
+})
+
+test_that("filtering by language does not fail when NULL is passed as input", {
+  expect_no_error({
+    filtered_repos <- github_testhost_priv$filter_repos_table_by_language(
+      repos_table = NULL,
+      language_filter = "R"
+    )
+  })
 })